Crate acvm_js

Modules

black_box_solvers 🔒

build_info 🔒

compression 🔒

execute 🔒

foreign_call 🔒

js_execution_error 🔒

js_witness_map 🔒

js_witness_stack 🔒

logging 🔒

public_witness 🔒

Structs

JsExecutionError

JsSolvedAndReturnWitness

JsWitnessMap

JsWitnessStack

Functions

and

Performs a bitwise AND operation between lhs and rhs

blake2s256

Calculates the Blake2s256 hash of the input bytes

build_info

Returns the BuildInfo object containing information about how the installed package was built. @returns {BuildInfo} - Information on how the installed package was built.

compress_witness

Compresses a WitnessMap into the binary format outputted by Nargo.

compress_witness_stack

Compresses a WitnessStack into the binary format outputted by Nargo.

decompress_witness

Decompresses a compressed witness as outputted by Nargo into a WitnessMap. This should be used to only fetch the witness map for the main function.

decompress_witness_stack

Decompresses a compressed witness stack as outputted by Nargo into a WitnessStack.

ecdsa_secp256k1_verify

Verifies a ECDSA signature over the secp256k1 curve.

ecdsa_secp256r1_verify

Verifies a ECDSA signature over the secp256r1 curve.

execute_circuit

Executes an ACIR circuit to generate the solved witness from the initial witness.

execute_circuit_with_return_witness

Executes an ACIR circuit to generate the solved witness from the initial witness. This method also extracts the public return values from the solved witness into its own return witness.

execute_program

Executes an ACIR circuit to generate the solved witness from the initial witness.

get_public_parameters_witness

Extracts a WitnessMap containing the witness indices corresponding to the circuit’s public parameters.

get_public_witness

Extracts a WitnessMap containing the witness indices corresponding to the circuit’s public inputs.

get_return_witness

Extracts a WitnessMap containing the witness indices corresponding to the circuit’s return values.

init_log_level

Sets the package’s logging level.

sha256_compression

Sha256 compression function

xor

Performs a bitwise XOR operation between lhs and rhs