acvm_js::execute
Function execute_circuit_with_return_witness

pub async fn execute_circuit_with_return_witness(
    program: Vec<u8>,
    initial_witness: JsWitnessMap,
    foreign_call_handler: ForeignCallHandler,
) -> Result<JsSolvedAndReturnWitness, Error>

Executes an ACIR circuit to generate the solved witness from the initial witness. This method also extracts the public return values from the solved witness into its own return witness.

@param {Uint8Array} circuit - A serialized representation of an ACIR circuit @param {WitnessMap} initial_witness - The initial witness map defining all of the inputs to circuit.. @param {ForeignCallHandler} foreign_call_handler - A callback to process any foreign calls from the circuit. @returns {SolvedAndReturnWitness} The solved witness calculated by executing the circuit on the provided inputs, as well as the return witness indices as specified by the circuit.