acvm_js
Module execute

Structs

ProgramExecutor 🔒

Functions

execute_circuit

Executes an ACIR circuit to generate the solved witness from the initial witness.

execute_circuit_pedantic 🔒

execute_circuit with pedantic ACVM solving

execute_circuit_with_return_witness

Executes an ACIR circuit to generate the solved witness from the initial witness. This method also extracts the public return values from the solved witness into its own return witness.

execute_circuit_with_return_witness_pedantic 🔒

executeCircuitWithReturnWitness with pedantic ACVM execution

execute_program

Executes an ACIR circuit to generate the solved witness from the initial witness.

execute_program_pedantic 🔒

execute_program with pedantic ACVM solving

execute_program_with_native_program_and_return 🔒

execute_program_with_native_type_return 🔒